Separate our SSH connector tests.
